\name{CutDat}
\alias{CutDat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
CutDat
}
\description{
 A rule to cut training and testing data
}
\usage{
CutDat
}
%- maybe also 'usage' for other objects documented here.


\references{
%%
}
\author{
 NingXin,WangQianQian,DataXujing
}
\note{

}

%% Make other sections like Warning with \section{Warning }{....}


\examples{
CutDat
}
